update bug_report template
We've seen a few cases in the past where the user accidentally leaked an access token or API key while reporting an issue. This puts the user at risk for an account take over or information disclosure. 

If we add the following line in our bug report template, we might prevent such leaks. 

`Please make sure the screenshot does not contain any API keys or access tokens.`